Work on reconciling backend + frontend physics
 * Fix some bugs in the frontend physics
   * Rotate entities around their centers of mass rather than their local origins
     * Send along `center_of_mass` in `CreationEvent` to make this possible
 * Move `set_position` into an `Entity` trait
 * Work towards properly handling spawning of the player entity to preserve fastpath
 * More message bugfixes on the backend for snapshot serialization and other things

+++ physics/src/physics/entities.rs
@@ -28,7 +28,7 @@ pub fn create_player_shape_handle(size: f32) -> ShapeHandle<f32> {
 
 pub enum EntityType {
     Player {
-        size: f32,
+        size: u32,
         movement: Movement,
         beam_aim: f32,
         beam_on: bool,
@@ -85,7 +85,7 @@ impl EntityType {
 
     pub fn get_shape_handle(&self) -> ShapeHandle<f32> {
         match self {
-            EntityType::Player { size, .. } => create_player_shape_handle(*size),
+            EntityType::Player { size, .. } => create_player_shape_handle(*size as f32),
             EntityType::Asteroid { vertices, .. } => {
                 let shape = ConvexPolygon::try_new(vertices.clone())
                     .expect("Unable to compute `ConvexPolygon` from asteroid vertices!");
